Velocity implemented. Not quite working

+++ Logic/IBallLogic.cs
@@ -2,8 +2,8 @@
 {
     public interface IBallLogic : IDisposable
     {
-        abstract int X { get; set; }
-        abstract int Y { get; set; }
+        abstract double X { get; set; }
+        abstract double Y { get; set; }
         abstract int Radius { get; }
 
         event Action<Object> PropertyChanged;
