+++ Model/IBallModel.cs
@@ -4,8 +4,8 @@ namespace Model
 {
     public interface IBallModel : INotifyPropertyChanged, IDisposable
     {
-        abstract int XHandler { get; set; }
-        abstract int YHandler { get; set; }
+        abstract double XHandler { get; set; }
+        abstract double YHandler { get; set; }
         abstract int RadiusHandler { get; }
 
         event PropertyChangedEventHandler? PropertyChanged;
